backend/requirements.txt
fastapi
uvicorn[standard] # Includes standard dependencies for uvicorn
pydantic
pydantic-settings   
sqlalchemy
psycopg2-binary
alembic
passlib[bcrypt]
bcrypt==4.0.1  # Pin to 4.0.1 for passlib compatibility (newer versions have breaking changes)
python-jose[cryptography]
python-multipart
google-genai